refactor: rename ecommerce components for medical domain
