+++ next-dashboard/src/components/patient/PresentingSymptoms.tsx
@@ -9,7 +9,7 @@ interface SymptomDetail {
   subDetail?: string;
 }
 
-const Presentations: React.FC = () => {
+const PresentingSymptoms: React.FC = () => {
   return (
     <div className="grid grid-cols-1 gap-4 md:grid-cols-2 md:gap-6">
       {patient.presenting_symptoms.map((symptom) => (
@@ -51,4 +51,4 @@ const Presentations: React.FC = () => {
   );
 };
 
-export default Presentations;
+export default PresentingSymptoms;
